api/requirements.txt
tensorflow==1.13.2
numpy==1.16.2
scikit-image==0.16.1
Pillow==6.2.0
Django==2.2.6
djangorestframework==3.10.3
uwsgi==2.0.18